# interpreter: removes unneeded fields from func instances