improves instantiation perf:

```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-10             34.2µs ± 5%    29.6µs ± 1%  -13.65%  (p=0.000 n=10+8)
Initialization/interpreter-multiple-10    13.7µs ± 3%    12.9µs ± 4%   -6.00%  (p=0.000 n=9+9)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              140kB ± 0%     138kB ± 0%   -1.62%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-10     141kB ± 0%     138kB ± 0%   -1.61%  (p=0.000 n=9+9)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10               93.0 ± 0%      52.0 ± 0%  -44.09%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-10      99.0 ± 0%      58.0 ± 0%  -41.41%  (p=0.000 n=9+9)
```
Signed-off-by: Takeshi Yoneda <takeshi@tetrate.io>